Se creo el metodo de generar aleatorio el serial para los estudiantes de ingenieria y se corrigieron algunos errores

+++ src/Validacion.java
@@ -154,14 +154,14 @@ public class Validacion {
     public String eleccionAValidarSinSobrepasar(int opcion, String cadena,int noSobrepasar) throws IOException {
         BufferedReader cp=new BufferedReader(new InputStreamReader(System.in));
 
-        int IsTrue;
-        IsTrue = Integer.parseInt(eleccionAValidar(opcion,cadena));
+        double IsTrue;
+        IsTrue = Double.parseDouble(eleccionAValidar(opcion,cadena));
 
         while (IsTrue>noSobrepasar){
             System.out.println("El dato no debe sobre pasar el numero "+noSobrepasar);
             System.out.println("ingrese el dato nuevamente: ");
             cadena=cp.readLine();
-            IsTrue = Integer.parseInt(eleccionAValidar(opcion,cadena ));
+            IsTrue = Double.parseDouble(eleccionAValidar(opcion,cadena ));
         }
 
         return cadena;
